engine/requirements.txt
playwright
beautifulsoup4
requests
python-dotenv
pocketbase
loguru
